reinit cap touch untouched now uses a moving average to avoid transients, fixed touchRead round robin behavior, now transmitting cap touch debug info, led brightness and led mode foundation laid, fixed switch ambiguity (ICCSwitch), better tx debug menu display, ability to reset tx stats, throttler staggering

+++ MAIN/CCSynth.h
@@ -113,7 +113,7 @@ struct VoiceList
     patch4(mix4, 0, mixOutp, 3),
     patchOut(mixOutp, 0, destMix, destPort)
   {
-    for (int i = 1; i < SizeofStaticArray(voices); ++ i) {
+    for (size_t i = 1; i < SizeofStaticArray(voices); ++ i) {
 //      // disable all but 1 layer
       voices[i].patchOut.disconnect();
     }
